requirements.txt
beautifulsoup4==4.13.3
config==0.5.1
google_api_python_client==2.156.0
google_auth_oauthlib==1.2.1
google_cloud_aiplatform==1.76.0
gspread==6.1.4
matplotlib==3.10.0
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
protobuf==5.29.3
pyecharts==2.0.7
Requests==2.32.3
snapshot_selenium==0.0.2
